+++ src/components/Card.svelte
@@ -23,7 +23,7 @@
   });
 </script>
 
-<div class="card cursor-pointer modal shadow-md">
+<div class="card modal shadow-md">
   <h3 class="modal-title">{title}</h3>
   <div class="flex flex-col h-[125px]">
     <div